// src/types/api.ts
export interface User {
	/** Unique stable id of the user */
	id: string;
	/** Name of the user */
	handle: string;
	/** URL link to the user's profile image */
	img_url: string;
	/** Email associated with the user's account. This will only be present on the /v1/me endpoint */
	email?: string;
}
// ...
export interface StyleMetadata {
	/** The unique identifier of the style */
	key: string;
	/** The unique identifier of the file which contains the style */
	file_key: string;
	/** Id of the style node within the figma file */
	node_id: string;
	/** The type of style */
	style_type: StyleType;
	/** URL link to the style's thumbnail image */
	thumbnail_url: string;
	/** Name of the style */
	name: string;
	/** The description of the style as entered by the publisher */
	description: string;
	/** The UTC ISO 8601 time at which the component set was created */
	created_at: string;
	/** The UTC ISO 8601 time at which the style was updated */
	updated_at: string;
	/** The user who last updated the style */
	sort_position: string;
	/** A user specified order number by which the style can be sorted */
	user: User;
}
// ...
export interface GetFileNodesResult {
	name: string;
	lastModified: string;
	thumbnailUrl: string;
	version: string;
	err?: string;
	nodes: {
		[nodeId: string]: {
			document: Node<"DOCUMENT">;
			components: { [nodeId: string]: Component };
			schemaVersion: number;
			styles: { [styleName: string]: Style };
		} | null;
	};
}
// ...
export interface GetFileStylesResult {
	status?: number;
	error?: Boolean;
	meta?: {
		styles: StyleMetadata[];
	};
}